requirements.txt
routeros_api
python-dotenv
wireguard
wgconfig
qrcode
image
markdown
jinja2
pyyaml
influxdb-client
icmplib
pandas
flask